Move `DashboardComponent` to its own file
